ajou_university/business_analytics_capstone_design/backend/requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
Django==4.0
django-cors-headers==3.10.1
django-environ==0.8.1
djangorestframework==3.12.4
google-api-core==2.3.0
google-auth==2.3.3
google-cloud-secret-manager==2.8.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
idna==3.3
libcst==0.3.23
mypy-extensions==0.4.3
mysqlclient==2.1.0
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-environ==0.4.54
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
typing-inspect==0.7.1
urllib3==1.26.7
